% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{getAWSTimeRange}
\alias{getAWSTimeRange}
\title{Get AWS start and end time.}
\usage{
getAWSTimeRange(id, network, aws_dir)
}
\arguments{
\item{id}{ID of the AWS.}

\item{network}{the AWS network code; 
1: tahmo, 2: campbell, 3: sutron, 4: seba, 5: microstep, 6: adcon.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/KMD_AWS_v2"}
}
\value{
a JSON object
}
\description{
Get the start and end time of a specified AWS.
}
