% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_winddata.R
\name{openairWindRose}
\alias{openairWindRose}
\title{Plot wind rose.}
\usage{
openairWindRose(net_aws, height, tstep, start, end, aws_dir)
}
\arguments{
\item{net_aws}{the network code and AWS ID, form <network code>_<AWS ID>. AWS network code, 1: tahmo, 2: campbell, 3: sutron, 4: seba, 5: microstep, 6: adcon.}

\item{height}{the observation height.}

\item{tstep}{time step of the data.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory containing ADT.\cr
Example: "D:/KMD_AWS_v2"}
}
\value{
Png image
}
\description{
Plot wind rose with openair for download.
}
